Crate objektdb_macros

Macros

delete_db

Allows you to delete a database by passing it the name of the database as a parameter.

Attribute Macros

objekt

It should be inserted on top of your structs to declare that it is an entity whose instances you want to store. The database in which you want to store the data should be specified. If the specified database does not exist, it is automatically created: so if you mistype the database name and enter one that does not exist, a new one will automatically be created. For the moment , since a language for interacting with data directly is not yet available, to delete a database you can make use of the macro delete_db!(“database_name”).

objekt_impl