objektdb_macros

Attribute Macro objekt

#[objekt]

It should be inserted on top of your structs to declare that it is an entity whose instances you want to store. The database in which you want to store the data should be specified. If the specified database does not exist, it is automatically created: so if you mistype the database name and enter one that does not exist, a new one will automatically be created. For the moment , since a language for interacting with data directly is not yet available, to delete a database you can make use of the macro delete_db!(“database_name”).

In addition, it also implements CRUD trait to perform transactions on the database. In particular, it implements the following functions:

select()

update()

where()

delete()

Example

use objektoDB::*;
#[objekt("my_database.db")]
struct Person {
   name: String,
   age: u32,
}